services/generative-tryon/app/services/tryon_pipeline.py
"""
Try-On Pipeline Service
Stable Diffusion + ControlNet for virtual try-on
"""

import uuid
from typing import Dict, Optional

import structlog
from PIL import Image

logger = structlog.get_logger()

class TryonPipeline:
    """Generative try-on pipeline using Stable Diffusion + ControlNet"""
    
    def __init__(self, settings):
        self.settings = settings
        self.models_loaded = False
        logger.info("tryon_pipeline.initialized")
    
    async def generate_tryon(
        self,
        job_id: str,
        profile_id: str,
        product_id: str,
        product_image_url: str,
        profile_photo_url: str,
        pose_image_url: Optional[str],
        control_type: str,
        num_inference_steps: int,
        guidance_scale: float
    ) -> Dict:
        """
        Generate virtual try-on image
        
        Steps:
        1. Download images (product, profile, pose)
        2. Segment product using SAM
        3. Extract pose using ControlNet
        4. Generate try-on using Stable Diffusion + ControlNet
        5. Upload result to S3
        """
        try:
            logger.info("tryon.generation.started", 
                      job_id=job_id, 
                      profile_id=profile_id,
                      product_id=product_id,
                      control_type=control_type)
            
            # Step 1: Download images (placeholder)
            product_image = await self._download_image(product_image_url)
            profile_image = await self._download_image(profile_photo_url)
            pose_image = await self._download_image(pose_image_url) if pose_image_url else profile_image
            
            # Step 2: Segment product using SAM (placeholder)
            product_mask = await self._segment_product(product_image)
            
            # Step 3: Extract pose using ControlNet (placeholder)
            pose_condition = await self._extract_pose(pose_image, control_type)
            
            # Step 4: Generate try-on (placeholder)
            result_image = await self._generate_inpainting(
                profile_image,
                product_image,
                product_mask,
                pose_condition,
                num_inference_steps,
                guidance_scale
            )
            
            # Step 5: Upload to S3 (placeholder)
            result_url = await self._upload_result(job_id, result_image)
            
            logger.info("tryon.generation.completed", job_id=job_id)
            
            return {
                "result_url": result_url,
                "job_id": job_id,
                "control_type": control_type
            }
            
        except Exception as e:
            logger.error("tryon.generation.failed", job_id=job_id, error=str(e))
            raise
    
    async def _download_image(self, url: str) -> Image.Image:
        """Download image from URL"""
        # In production, use httpx to download
        # For demo, return a placeholder
        return Image.new('RGB', (512, 512), color='white')
    
    async def _segment_product(self, product_image: Image.Image) -> Image.Image:
        """Segment product from background using SAM"""
        # In production, load SAM model and segment
        # For demo, return a placeholder mask
        return Image.new('L', (512, 512), color=255)
    
    async def _extract_pose(self, image: Image.Image, control_type: str) -> Dict:
        """Extract pose/depth/edge condition using ControlNet"""
        # In production, load ControlNet model and extract condition
        # For demo, return placeholder
        return {
            "type": control_type,
            "condition": "placeholder"
        }
    
    async def _generate_inpainting(
        self,
        profile_image: Image.Image,
        product_image: Image.Image,
        product_mask: Image.Image,
        pose_condition: Dict,
        num_inference_steps: int,
        guidance_scale: float
    ) -> Image.Image:
        """Generate try-on using Stable Diffusion + ControlNet inpainting"""
        # In production, load Stable Diffusion + ControlNet models
        # For demo, return a placeholder
        return Image.new('RGB', (512, 512), color='blue')
    
    async def _upload_result(self, job_id: str, image: Image.Image) -> str:
        """Upload result to S3"""
        # In production, upload to S3
        return f"s3://{self.settings.s3_bucket}/tryon-results/{job_id}.png"
