services/generative-tryon/requirements.txt
# DS3 Generative Try-On Service
# Task 3.2: Generative Try-On Pipeline

# Web Framework
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart>=0.0.22

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25

# ML/AI - Stable Diffusion
torch==2.6.0
torchvision==0.21.0
diffusers==0.25.0
transformers==4.37.0
accelerate==0.26.1
safetensors==0.4.3

# ControlNet
controlnet-aux==0.0.7

# SAM (Segment Anything)
segment-anything==1.0

# Image Processing
Pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.3

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Logging & Monitoring
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
